user profile & protect content & webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 /*
 export const home = (req, res) => {
     Video.find({}, (error, videos) => {
@@ -19,7 +20,8 @@ export const home = async(req, res) => {
 }
 export const watch = async(req, res) => {
     const {id} = req.params;
-    const video = await Video.findById(id);
+    const video = await (await Video.findById(id)).populate("owner");
+    console.log(video);
     if(video === null){
         return res.render("404", {pageTitle:"Video not found."});
     };
@@ -27,19 +29,26 @@ export const watch = async(req, res) => {
 }
 export const getEdit = async(req, res) => {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});   
     };
-
+    if(String(video.owner) !== _id){
+        return res.status(403).redirect("/");
+    }
     return res.render("edit", {pageTitle: `Edit ${video.title}`, video});
 }
 export const postEdit = async(req, res) => {
+    const {user: {_id}}=req.session;
     const {id} = req.params;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({_id: id});
+    const video = await Video.findOne({_id: id});
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== _id){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {title,
                                  description,
@@ -50,15 +59,20 @@ export const getUpload = (req, res) => {
     return res.render("upload", {pageTitle: "Upload Video"});
 };
 export const postUpload = async(req, res) => {
-    const {path:fileUrl} = req.file
-    const {body: {title, description, hashtags}} = req;
+    const {_id} = req.session.user;
+    const {title, description, hashtags} = req.body;
+    const {path:fileUrl} = req.file;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             fileUrl,
             description,
             hashtags: Video.formatHashtags(hashtags),
+            owner: _id,
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        await user.save()
     } catch(error) {
         return res.status(400).render("upload", {pageTitle: "Upload Video", errorMessage: error._message});
     }
@@ -66,11 +80,15 @@ export const postUpload = async(req, res) => {
 };
 
 export const deleteVideo = async(req, res) => {
+    const {user: {_id}} = req.session;
     const {params:{id}} = req;
-    const video = await Video.exists({_id:id});
+    const video = await Video.findOne({_id:id});
     if(!video){
         return res.render("404", {pageTitle: "Video not found."});
     }
+    if(String(video.owner) !== _id){
+        return res.status(304).redirect("/");
+    };
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
